requirements.txt
streamlit
openai>=1.3.0
langchain>=0.1.17
faiss-cpu
python-dotenv
PyMuPDF